Extract items list component logic for re-usability
