+++ client/my-sites/plans/jetpack-plans/product-store/items-list/all-items.tsx
@@ -1,0 +1,77 @@
+import classNames from 'classnames';
+import { useStoreItemInfo } from '../hooks/use-store-item-info';
+import SimpleProductCard from '../simple-product-card';
+import { MostPopularProps } from '../types';
+
+import './style.scss';
+
+export const AllItems: React.FC< MostPopularProps > = ( {
+	className,
+	createCheckoutURL,
+	duration,
+	heading,
+	items,
+	onClickMoreInfoFactory,
+	onClickPurchase,
+	siteId,
+} ) => {
+	const {
+		getCheckoutURL,
+		getCtaLabel,
+		getOnClickPurchase,
+		isIncludedInPlan,
+		isIncludedInPlanOrSuperseded,
+		isOwned,
+		isPlanFeature,
+		isSuperseded,
+		isDeprecated,
+		isUserPurchaseOwner,
+	} = useStoreItemInfo( {
+		createCheckoutURL,
+		onClickPurchase,
+		duration,
+		siteId,
+	} );
+
+	const wrapperClassName = classNames( 'jetpack-product-store__all-items', className );
+
+	return (
+		<div className={ wrapperClassName }>
+			<h3 className="jetpack-product-store__all-items--header">{ heading }</h3>
+
+			<div className="jetpack-product-store__all-items--grid">
+				{ items.map( ( item ) => {
+					const isItemOwned = isOwned( item );
+					const isItemSuperseded = isSuperseded( item );
+					const isItemDeprecated = isDeprecated( item );
+					const isItemIncludedInPlanOrSuperseded = isIncludedInPlanOrSuperseded( item );
+
+					const isCtaDisabled =
+						( isItemOwned || isIncludedInPlan( item ) ) && ! isUserPurchaseOwner( item );
+
+					const ctaLabel = getCtaLabel( item );
+
+					const hideMoreInfoLink =
+						isItemDeprecated || isItemOwned || isItemIncludedInPlanOrSuperseded;
+
+					return (
+						<SimpleProductCard
+							checkoutURL={ getCheckoutURL( item ) }
+							ctaAsPrimary={ ! ( isItemOwned || isPlanFeature( item ) || isItemSuperseded ) }
+							ctaLabel={ ctaLabel }
+							isCtaDisabled={ isCtaDisabled }
+							isIncludedInPlan={ isItemIncludedInPlanOrSuperseded }
+							hideMoreInfoLink={ hideMoreInfoLink }
+							isOwned={ isItemOwned }
+							item={ item }
+							key={ item.productSlug }
+							onClickMore={ onClickMoreInfoFactory( item ) }
+							onClickPurchase={ getOnClickPurchase( item ) }
+							siteId={ siteId }
+						/>
+					);
+				} ) }
+			</div>
+		</div>
+	);
+};
