+++ client/my-sites/plans/jetpack-plans/product-store/items-list/style-most-popular.scss
@@ -30,3 +30,29 @@
 		gap: 24px;
 	}
 }
+
+.jetpack-product-store__most-popular--item {
+	display: flex;
+	flex-direction: column;
+}
+
+.jetpack-product-store__most-popular--item .foldable-card {
+	box-shadow: none;
+	width: 100%;
+	margin: 0;
+	border-bottom: 1px solid var( --studio-gray-5 );
+}
+
+.jetpack-product-store__most-popular--item .foldable-card.is-expanded {
+	margin: 0;
+}
+
+.jetpack-product-store__most-popular--item .foldable-card.is-expanded .foldable-card__content {
+	padding: 0;
+}
+
+.jetpack-product-store__most-popular--item .foldable-card__main {
+	font-size: $font-body;
+	line-height: 24px;
+	font-weight: 600;
+}
