+++ client/my-sites/plans/jetpack-plans/product-store/items-list/style.scss
@@ -1,16 +1,32 @@
 @import '@wordpress/base-styles/breakpoints';
 @import '@wordpress/base-styles/mixins';
 
-.jetpack-product-store__products-list-all {
+.jetpack-product-store__items-list .button.is-primary {
+	color: var( --color-text-inverted );
+	min-width: 5.7em;
+	border: none;
+}
+
+.jetpack-product-store__all-items--grid {
+	display: grid;
+	column-gap: 80px;
+	row-gap: 32px;
+
+	@include break-medium {
+		grid-template-columns: repeat( 2, 1fr );
+	}
+}
+
+.jetpack-product-store__all-items {
 	margin-top: 4em;
 	margin-bottom: 48px;
 }
 
-.jetpack-product-store__products-list-all h3 {
+.jetpack-product-store__all-items h3 {
 	margin-bottom: 2.5em;
 }
 
-.jetpack-product-store__products-list-all-header {
+.jetpack-product-store__all-items--header {
 	margin-top: rem( 32px );
 	margin-bottom: rem( 24px );
 	font-size: $font-title-small;
@@ -22,15 +38,5 @@
 		font-size: $font-title-medium;
 		line-height: rem( 29px );
 		font-weight: 600;
-	}
-}
-
-.jetpack-product-store__products-list-all-grid {
-	display: grid;
-	column-gap: 80px;
-	row-gap: 32px;
-
-	@include break-medium {
-		grid-template-columns: repeat( 2, 1fr );
 	}
 }
