+++ client/my-sites/plans/jetpack-plans/product-store/types.ts
@@ -53,10 +53,11 @@ export interface ItemsListProps extends ItemToDisplayProps {
 	currentView: ViewType;
 }
 
-export type MostPopularProps = {
+export type MostPopularProps = UseStoreItemInfoProps & {
 	className?: string;
 	heading: string;
-	items: React.ReactNode;
+	items: Array< SelectorProduct >;
+	onClickMoreInfoFactory: ( item: SelectorProduct ) => VoidFunction;
 };
 
 export type HeroImageProps = {
